+++ tests/e2e/test_routing/test_path_mounting.py
@@ -1,12 +1,8 @@
-import subprocess
-import time
 from pathlib import Path
 from typing import TYPE_CHECKING, Callable, List
 
 import httpx
-import psutil
 import pytest
-from _pytest.fixtures import FixtureRequest
 from _pytest.monkeypatch import MonkeyPatch
 
 from litestar import Litestar, MediaType, asgi, get, websocket
@@ -101,30 +97,6 @@ def test_does_not_support_asgi_handlers_on_same_level_as_websockets() -> None:
         Litestar(route_handlers=[asgi_handler, regular_handler])
 
 
-@pytest.fixture()
-def run_server(request: FixtureRequest) -> Callable[[List[str]], None]:
-    def runner(server_command: List[str]) -> None:
-        proc = psutil.Popen(server_command, stderr=subprocess.PIPE, stdout=subprocess.PIPE)
-
-        def kill() -> None:
-            for child in proc.children(recursive=True):
-                child.kill()
-            proc.kill()
-
-        request.addfinalizer(kill)
-
-        for _ in range(10):
-            try:
-                httpx.get("http://127.0.0.1:9999/", timeout=0.1)
-                break
-            except httpx.TransportError:
-                time.sleep(0.1)
-        else:
-            raise RuntimeError("App failed to come online")
-
-    return runner
-
-
 @pytest.mark.parametrize(
     "server_command",
     [
@@ -134,13 +106,11 @@ def run_server(request: FixtureRequest) -> Callable[[List[str]], None]:
     ],
 )
 @pytest.mark.xdist_group("live_server_test")
+@pytest.mark.server_integration
 def test_path_mounting_live_server(
-    tmp_path: Path, monkeypatch: MonkeyPatch, server_command: List[str], run_server: Callable[[List[str]], None]
+    tmp_path: Path, monkeypatch: MonkeyPatch, server_command: List[str], run_server: Callable[[str, List[str]], None]
 ) -> None:
-    # https://github.com/litestar-org/litestar/issues/2998
-    app_file = tmp_path / "app.py"
-    app_file.write_text(
-        """
+    app = """
 from litestar import asgi, Litestar
 from litestar.types import Receive, Scope, Send
 from litestar.response.base import ASGIResponse
@@ -152,10 +122,8 @@ async def handler(scope: Scope, receive: Receive, send: Send) -> None:
 
 
 app = Litestar(route_handlers=[handler])
-    """
-    )
-    monkeypatch.chdir(tmp_path)
-    run_server(server_command)
+"""
+    run_server(app, server_command)
 
     res = httpx.get("http://127.0.0.1:9999/sub/path/fragment")
     assert res.status_code == 200
